pyproject.toml
[tool.poetry]
name = "2018-11-16-bf-api-ibge"
version = "0.1.0"
description = "Extraí e padroniza as apis do IBGE em https://servicodados.ibge.gov.br"
authors = [ "Beta Faccion <betafcc@gmail.com>" ]

[tool.poetry.dependencies]
python = "^3.7"
tornado = "<5"
ipython = "^7.1"
jupyterlab = "^0.35.3"
requests = "^2.20"
parsel = "^1.5"
sh = "^1.12"
oyaml = "^0.7.0"
python-box = "^3.2"

[tool.poetry.dev-dependencies]
pytest = "^3.9"
flake8 = "^3.6"
mypy = "^0.641.0"

[tool.pandora]
name = "Api IBGE"
date = 2018-11-16T11:14:25.000Z

[build-system]
requires = [ "poetry>=0.12" ]
build-backend = "poetry.masonry.api"
